Split off EnumValue into it's own file.
